Rename hints pkg to intents
