+++ xicam/XPCS/intents/canvases.py
